--- a/apps/main-site/src/app/globals.css
+++ b/apps/main-site/src/app/globals.css
@@ -8,8 +8,9 @@ html,
 body {
 	max-width: 100vw;
 	overflow-x: hidden;
-	font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell,
-		"Open Sans", "Helvetica Neue", sans-serif;
+	font-family:
+		-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans",
+		"Helvetica Neue", sans-serif;
 }
 
 a {