--- a/apps/product/.reassure/config.ts
+++ b/apps/product/.reassure/config.ts
@@ -2,7 +2,7 @@ import { configure } from "reassure";
 
 configure({
 	testCommand:
-		'pnpm test --testMatch="**/*.perf.test.tsx" --testRunner "apps/product/jest-perf.config.js"',
+		'pnpm test --testMatch="**/*.perf.test.tsx" --config "apps/product/jest-perf.config.js"',
 	// Run performance tests in a more isolated environment
 	verbose: true,
 	// Increase the number of measurements for more accurate results