--- a/apps/product/package.json
+++ b/apps/product/package.json
@@ -11,6 +11,9 @@
 		"dev": "expo start",
 		"build": "expo export",
 		"test": "jest --passWithNoTests",
+		"test:perf": "jest --testMatch='**/*.perf.test.tsx' --config=jest-perf.config.js",
+		"reassure": "reassure",
+		"reassure:baseline": "reassure --baseline",
 		"lint": "biome check --fix --vcs-root=../.. src",
 		"typecheck": "tsc --noEmit",
 		"clean": "rm -rf .expo dist web-build",
@@ -28,6 +31,8 @@
 		"@react-native-community/netinfo": "^11.4.1",
 		"@react-navigation/bottom-tabs": "^7.0.0",
 		"@react-navigation/native": "^7.0.0",
+		"@tanstack/react-query": "^5.21.1",
+		"@tanstack/react-query-persist-client": "^5.21.1",
 		"expo": "~53.0.12",
 		"expo-application": "^6.1.4",
 		"expo-blur": "~14.1.5",
@@ -52,9 +57,7 @@
 		"react-native-svg": "15.11.2",
 		"react-native-web": "~0.20.0",
 		"react-native-webview": "13.13.5",
-		"sentry-expo": "^7.2.0",
-		"@tanstack/react-query": "^5.21.1",
-		"@tanstack/react-query-persist-client": "^5.21.1"
+		"sentry-expo": "^7.2.0"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.25.2",
@@ -67,6 +70,7 @@
 		"jest": "~29.7.0",
 		"jest-expo": "~53.0.7",
 		"playwright": "^1.53.1",
-		"react-test-renderer": "19.0.0"
+		"react-test-renderer": "19.0.0",
+		"reassure": "^1.4.0"
 	}
 }