## Summary
Implements automated performance regression testing for React Native components using Reassure.

## What is Reassure?
Reassure measures React component render performance and detects regressions by comparing:
- **Render Duration**: How long components take to render
- **Render Count**: How many times components re-render

## Implementation Details

### Performance Tests Added
- **Components**: Button, TextInput (critical BGUI components)  
- **Screens**: HomeScreen (example of testing full screens)
- **Location**: `apps/product/__tests__/**/*.perf.test.tsx`

### GitHub Actions Integration
- New workflow: `.github/workflows/performance.yml`
- Triggers on PRs that modify `apps/product` or `packages/bgui`
- Compares PR branch performance against base branch
- Posts results as PR comment

### Local Development
```bash
# From root directory
pnpm perf           # Run performance comparison
pnpm perf:baseline  # Generate baseline measurements

# Or use the helper script
./scripts/run-perf-tests.sh [baseline|compare|test]
```

### Configuration
- Test runner: `jest-perf.config.js` (optimized for performance testing)
- Reassure config: `.reassure/config.ts` (20 runs, 5 warmup runs)
- Documentation: `docs/REASSURE_PERFORMANCE.md`

## Benefits
1. **Catch Performance Regressions**: Before they reach production
2. **Automated Testing**: No manual performance profiling needed
3. **PR Integration**: See performance impact directly in PRs
4. **Baseline Comparison**: Track performance over time

## Example PR Comment
```
📱 React Native Performance Results

Component    | Render Duration | Render Count | Status
-------------|-----------------|--------------|--------
Button       | 12ms → 13ms     | 1 → 1        | ✅ +8%
TextInput    | 18ms → 25ms     | 2 → 3        | ⚠️ +39%
HomeScreen   | 45ms → 42ms     | 3 → 3        | ✅ -7%
```

This complements the existing Lighthouse CI for web performance, giving us full performance coverage across all platforms.